Submit Button
Submit button now reads the contents of the textarea on the collection form. and displays them on the page, just to text getting the data from the box.

I had to turn off  EnableEventValidation on the CollectionEntry page to keep the button click from causing a postback error, wasn't able to figure out why that was happening. But for now, it works.