Added Functionality for Getting Username After Login
I set up a test by generating a cookie to the username of whoever logged in. We can convert this to a string and potentially call it from other parts of the project. The cookie by default expires when we close the browser. TODO: Work on encryption methods for both this cookie and storing passwords in UserInfo, C# has some built in libraries.